Download dataset and spacy language models
Added _download() function that downloads a file from the url and
saves the file to the desired path.
Added download() function that uses _download() to download the
datasets, SQuAD 2.0 and GLoVe 300D, from web and saves then to
the data directory. The function also downloads spacy english
language model which would be required as tokenizer during
pre-processing.