Added BiDAF Attention layer and BiDAF Output layer
The BiDAF Attention layers performs Context2Query and Query2Context
attention. This allows quetion and context words to interact with
each other and highlight words that are most relevant to the question
in context and vice-versa.
The BiDAF Output layer returns the probability of each word being
and end for the answer.